Ensure all split match files are created
The ad hoc relative finder pipeline expected all match files to exist
even if there were no matches in them. It's not clear if this is also
true for relative finder in production. This change opens all output
files before matches are written to them to ensure that all match files
are created.